sms-agent/pyproject.toml
[project]
name = "sms-agent"
version = "0.1.0"
description = "SMS Agent that responds to incoming messages using AI"
requires-python = ">=3.10,<3.14"
dependencies = [
    "livekit-agents>=1.3.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]
